+++ Liquibook.NET/Events/OnFillEventArgs.cs
@@ -8,11 +8,11 @@ namespace Liquibook.NET.Events
         public IOrder Order { get; }
         public IOrder MatchedOrder { get; }
         public Quantity FillQuantity { get; }
-        public int FillCost { get; }
+        public decimal FillCost { get; }
         public bool InboundOrderFilled { get; }
         public bool MatchedOrderFilled { get; }
 
-        public OnFillEventArgs(IOrder order, IOrder matchedOrder, Quantity fillQuantity, int fillCost,
+        public OnFillEventArgs(IOrder order, IOrder matchedOrder, Quantity fillQuantity, decimal fillCost,
             bool inboundOrderFilled, bool matchedOrderFilled)
         {
             Order = order;
